Ender 3 V2 with BLTouch

+++ Marlin/src/core/serial.h
@@ -299,7 +299,7 @@ void serial_echopair_PGM(PGM_P const s_P, unsigned char v);
 void serial_echopair_PGM(PGM_P const s_P, unsigned int v);
 void serial_echopair_PGM(PGM_P const s_P, unsigned long v);
 inline void serial_echopair_PGM(PGM_P const s_P, bool v)    { serial_echopair_PGM(s_P, (int)v); }
-inline void serial_echopair_PGM(PGM_P const s_P, void *v)   { serial_echopair_PGM(s_P, (uintptr_t)v); }
+inline void serial_echopair_PGM(PGM_P const s_P, void *v)   { serial_echopair_PGM(s_P, (unsigned long)v); }
 
 void serial_echo_start();
 void serial_error_start();
